requirements.txt
psycopg2==2.8.4
Django==3.0.3
djangorestframework==3.11.0
